+++ TPA_Exercises/src/lista2/ex8/profiles/Profile.java
@@ -8,5 +8,5 @@ import lista2.ex8.documents.Document;
  * @time 7:38:30 PM
  */
 public interface Profile {
-    public int compare(Document a, Document b);
+    public double compare(Document a, Document b);
 }
